Create standalone master module.
(index -> lib/master).js
	-Split the main module into another, `master.js`, which contains
	all functions related to managing `lib/worker.js` workers.
